requirements.txt
Flask[async]
pyyaml
flask_babel
httpx
requests
aiohttp
pandas
mwoauth